Refactor and update file names.
